+++ src/app/Scenes/MyProfile/__tests__/MyProfilePushNotifications.tests.tsx
@@ -1,11 +1,9 @@
 import { Switch, Text } from "@artsy/palette-mobile"
 import { fireEvent, screen } from "@testing-library/react-native"
 import { MyProfilePushNotificationsTestQuery } from "__generated__/MyProfilePushNotificationsTestQuery.graphql"
-import { updateMyUserProfileMutation } from "__generated__/updateMyUserProfileMutation.graphql"
 import { SwitchMenu } from "app/Components/SwitchMenu"
 import {
   MyProfilePushNotifications,
-  NOTIFICATION_SETTINGS,
   UserPushNotificationSettings,
 } from "app/Scenes/MyProfile/MyProfilePushNotifications"
 import { PushAuthorizationStatus } from "app/utils/PushNotification"
@@ -185,24 +183,6 @@ describe("MyProfilePushNotificationsQueryRenderer", () => {
           },
         })
       )
-    })
-  })
-  // This is a pure type test that ensures all notification keys exist in the mutation response type
-  it("mutation response includes all UserPushNotificationSettings fields from ProfilePushNotifications", () => {
-    // Import the mutation type
-    type MutationResponse = updateMyUserProfileMutation["response"]
-    type UpdateMyUserProfile = NonNullable<MutationResponse["updateMyUserProfile"]>
-    type MeObject = NonNullable<UpdateMyUserProfile["me"]>
-
-    // TypeScript will error here if any UserPushNotificationSettings key is not in MeObject
-    const _typeCheck = (k: UserPushNotificationSettings): keyof MeObject => {
-      return k // If this passes type-checking, the key exists on MeObject
-    }
-
-    // For each notification setting key, it must exist as a property on the Me object
-    // TypeScript compiler will catch any missing properties
-    Object.values(NOTIFICATION_SETTINGS).forEach((notification) => {
-      _typeCheck(notification.key)
     })
   })
 })
